requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
numpy==2.0.2
ta==0.11.0
ccxt==4.3.83
python-dotenv==1.0.1
scikit-learn==1.5.2
joblib==1.4.2
polars==1.12.0
schedule==1.2.2
psutil==6.1.0
httpx==0.27.2
requests==2.32.3
cachetools==5.5.0
pytz==2024.2
python-telegram-bot==21.4
